generators: Add copy generator
Closes #110

Copy any file from the host to the container using "destination".
Logic is very similar to that one from Docker.

First check if the input is a file or a directory.
Then check whether the destination finishes in a "/" or not
Afterwards, the rules for copying can be applied
Rules for copying:
 - If <src> is a directory, the entire contents of the directory
are copied. Only symlinks and regular files are supported.
Note 1: The directory itself is not copied, just its contents.
Note 2: For files copied, only regular unix permissions are kept
 - If <src> is any other kind of file, it is copied individually
along with its metadata. In this case, if <dest> ends with a trailing
slash /, it will be considered a directory and the contents of <src>
will be written at <dest>/base(<src>).
 - If <dest> does not end with a trailing slash, it will be considered
a regular file and the contents of <src> will be written at <dest>.
- If <dest> doesn’t exist, it is created along with all missing
directories in its path.
 - Multiple <src> resources can be specified using golang regexps.
For simplicity they are only allowed in the basename and not in
the directory hierarchy. If more than one match is found, <dest>
will be automatically interpreted as a directory

Signed-off-by: Pablo Correa Gómez <ablocorrea@hotmail.com>