Don't attempt to re-install packages that were installed early
If a package set has the `early: true` option set, those packages are
installed with the source (when using `debootstrap`). But because the
early packages remain in the set of packages, they are attempted to be
installed again.

This is a minor cosmetic issue, but not attempting to re-install the
early packages makes the install log slightly cleaner.

Also remove references to the old `early_packages` option in the docs.

Signed-off-by: Eddy Gurney <github@eddyg.promessage.com>