go.mod for building without GOPATH
This was mostly automatic except I had to add the entry for
github.com/mudler/docker-companion since they haven't cut a
release yet with a fix for an incorrect upstream import path.

This is enough that you can "make" as well as "go install
./distrobuilder/", but there may be other changes to the Makefile we
want to do in addition to this once everyone is off the old go versions.

Signed-off-by: Joel Jensen <yobert@gmail.com>