*: Switch to errors without stack traces
The zap logger includes the verbose output if the error implements the
Formatter interface which errors.Error (from the pkg/errors package)
does. This includes the stack trace which causes the visible error
message to be rather messy.

Using fmt.Errorf and errors.New (standard library) solves the problem
as these don't implement the Formatter interface.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>