udev: Allow NetworkManager to manage veth interfaces
NetworkManager ships with an udev rule that excludes veth devices.
The reason is that veths are often used by tooling that would be unhappy
if NetworkManager interfered.

In containers where /sys is read-only (such as libvirt's LXC backend),
systemd-udev is not used and NetworkManager is happy to manage them.
With LXD, however, /sys is mounted read-write and systemd-udev runs.

While some future version of NetworkManager might solve this somehow,
we need this to work with existing versions. Let's just override
the rule in question with something that works well with LXD.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>