main: Fix cache directory removal if it's a mountpoint
This fixes an issue where the cache directory cannot be removed if it's
a mountpoint.

This commit tries removing the cache directory but doesn't fail if
unsuccessful. Instead it tries removing the content of the cache
directory. Should this fail as well, distrobuilder will issue a warning.
The user is then responsible for cleaning up the cache directory.

Fixes #716

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>