main: Keep track of current subcommand
When app.Execute() fails, there's no way of knowing which command
failed. However, we want to know this as we call postRun in case of an
error, and handle the "validate" subcommand differently in that
function.

This commit adds a subCommand field to the global struct which is used
to keep track of the current subcommand.

Signed-off-by: Thomas Hipp <thomashipp@gmail.com>