sources: Alpaquita - use source.variant not image.variant
The image.variant is an arbitrary user choice (e.g. an image that
includes cloud-init), while possible source.variant values are
predefined by the distro (e.g. libc, as is the case with Alpaquita).

Unfortunately parts of distrobuilder code base are confused about
this, so I followed the wrong example when I first wrote this code.

Signed-off-by: Valery Ushakov <valery.ushakov@bell-sw.com>